pyproject.toml
[project]
name = "finaltask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.22",
    "django>=6.0",
    "django-ninja>=1.5.1",
    "django-storages>=1.14.6",
    "dockerfile>=3.4.0",
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
]
